pyproject.toml
[tool.poetry]
name = "snn_hfo_detection"
version = "0.1.0"
description = "HFO detection with a Spiking Neural Network"
authors = ["Karla Burelo <kburel@ini.uzh.ch>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
numpy = "^1.21.1"
Brian2 = "^2.4.2"
scipy = "^1.7.0"
seaborn = "^0.11.1"
teili = { git="https://gitlab.com/neuroinf/teili.git", branch="dev" }

[tool.poetry.dev-dependencies]
pytest-xdist = "^2.3.0"
pylint = "^2.9.5"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
